// src/types/enums.tsx
/* allerts enums */
export enum vertical {
  top = 'top',
  bottom = 'bottom'
};

export enum horizontal {
  left = 'left',
  center = 'center',
  right = 'right'
};

export enum type {
  success = 'success',
  error = 'error',
  info = 'info',
  warning = 'warning'
};

/* buttons */
export enum typesButtons {
  button = 'button',
  submit = 'submit',
  reset = 'reset'
}